pyproject.toml
# =============================================================================
# MAPLE-FEP: Enhanced pyproject.toml for PyPI Publishing
# =============================================================================
# This is an improved version of pyproject.toml for the new package repository.
# Key improvements:
# - Stricter dependency version bounds
# - Better classifiers for discoverability
# - Optional CLI entry point
# - Improved tool configurations
# - GitHub Actions release workflow integration
# =============================================================================

[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "maple-fep"
version = "0.1.0"
description = "Maximum A Posteriori Learning of Energies - Bayesian inference tools for Free Energy Perturbation (FEP) calculations in computational drug discovery"
readme = "README.md"
license = {text = "MIT"}
authors = [
    {name = "Aakankschit Nandkeolyar", email = "aakankschit@example.com"}
]
maintainers = [
    {name = "Aakankschit Nandkeolyar", email = "aakankschit@example.com"}
]
keywords = [
    "free energy perturbation",
    "FEP",
    "bayesian inference",
    "molecular dynamics",
    "computational chemistry",
    "drug discovery",
    "binding affinity",
    "thermodynamics",
    "cycle closure",
    "variational inference",
]
classifiers = [
    # Development Status
    "Development Status :: 3 - Alpha",

    # Intended Audience
    "Intended Audience :: Science/Research",
    "Intended Audience :: Developers",

    # Topic
    "Topic :: Scientific/Engineering :: Chemistry",
    "Topic :: Scientific/Engineering :: Bio-Informatics",
    "Topic :: Scientific/Engineering :: Artificial Intelligence",
    "Topic :: Scientific/Engineering :: Physics",

    # License
    "License :: OSI Approved :: MIT License",

    # Python versions
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3 :: Only",

    # Operating System
    "Operating System :: OS Independent",

    # Typing
    "Typing :: Typed",
]
requires-python = ">=3.10"

dependencies = [
    # Core scientific computing
    "numpy>=1.21.0,<3.0",
    "scipy>=1.7.0",
    "pandas>=1.3.0",

    # Deep learning / probabilistic programming
    "torch>=2.0.0",
    "pyro-ppl>=1.8.0",

    # Graph analysis
    "networkx>=2.6.0",

    # ML utilities
    "scikit-learn>=1.0.0",

    # Configuration / validation
    "pydantic>=2.0.0",

    # Visualization
    "matplotlib>=3.4.0",
    "seaborn>=0.11.0",

    # Utilities
    "tqdm>=4.60.0",
    "adjustText>=0.8",
    "requests>=2.25.0",
]

[project.optional-dependencies]
# Development dependencies
dev = [
    "pytest>=7.0",
    "pytest-cov>=4.0",
    "pytest-xdist>=3.0",  # Parallel test execution
    "black>=23.0",
    "ruff>=0.1.0",  # Fast linter (replaces flake8, isort)
    "mypy>=1.0",
    "pre-commit>=3.0",
    "ipython>=8.0",
]

# Documentation dependencies
docs = [
    "sphinx>=6.0",
    "sphinx-rtd-theme>=1.0",
    "sphinx-autodoc-typehints>=1.20",
    "sphinx-copybutton>=0.5",
    "myst-parser>=1.0",
    "nbsphinx>=0.9",  # Jupyter notebook integration
    "sphinx-autobuild>=2021.0",
]

# Jupyter/notebook dependencies
notebook = [
    "jupyter>=1.0",
    "jupyterlab>=4.0",
    "ipywidgets>=8.0",
]

# All optional dependencies
all = [
    "maple-fep[dev,docs,notebook]",
]

[project.urls]
Homepage = "https://github.com/aakankschit/maple-fep"
Documentation = "https://maple-fep.readthedocs.io/"
Repository = "https://github.com/aakankschit/maple-fep.git"
Changelog = "https://github.com/aakankschit/maple-fep/blob/main/CHANGELOG.md"
"Bug Tracker" = "https://github.com/aakankschit/maple-fep/issues"

# Optional: CLI entry point (uncomment if you want a command-line tool)
# [project.scripts]
# maple = "maple.cli:main"

# =============================================================================
# Setuptools configuration
# =============================================================================

[tool.setuptools.packages.find]
where = ["src"]
include = ["maple*"]

[tool.setuptools.package-dir]
"" = "src"

[tool.setuptools.package-data]
maple = ["py.typed"]  # PEP 561 marker for typed package

# =============================================================================
# Black - Code formatter
# =============================================================================

[tool.black]
line-length = 88
target-version = ['py310', 'py311', 'py312']
include = '\.pyi?$'
extend-exclude = '''
/(
    \.eggs
    | \.git
    | \.hg
    | \.mypy_cache
    | \.tox
    | \.venv
    | _build
    | buck-out
    | build
    | dist
    | docs/_build
)/
'''

# =============================================================================
# Ruff - Fast linter (replaces flake8, isort, and more)
# =============================================================================

[tool.ruff]
line-length = 88
target-version = "py310"

[tool.ruff.lint]
select = [
    "E",      # pycodestyle errors
    "W",      # pycodestyle warnings
    "F",      # Pyflakes
    "I",      # isort
    "B",      # flake8-bugbear
    "C4",     # flake8-comprehensions
    "UP",     # pyupgrade
    "ARG",    # flake8-unused-arguments
    "SIM",    # flake8-simplify
]
ignore = [
    "E501",   # line too long (handled by black)
    "B008",   # do not perform function calls in argument defaults
    "C901",   # too complex
    "ARG001", # unused function argument (common in Pyro models)
]

[tool.ruff.lint.per-file-ignores]
"tests/*" = ["ARG"]  # Allow unused arguments in tests
"__init__.py" = ["F401"]  # Allow unused imports in __init__.py

[tool.ruff.lint.isort]
known-first-party = ["maple"]

# =============================================================================
# Pytest configuration
# =============================================================================

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = "test_*.py"
python_classes = "Test*"
python_functions = "test_*"
addopts = [
    "-v",
    "--tb=short",
    "--strict-markers",
    "-ra",  # Show summary of all test outcomes
]
markers = [
    "slow: marks tests as slow (deselect with '-m \"not slow\"')",
    "integration: marks tests as integration tests",
    "gpu: marks tests requiring GPU",
]
filterwarnings = [
    "ignore::DeprecationWarning",
    "ignore::PendingDeprecationWarning",
]

# =============================================================================
# Coverage configuration
# =============================================================================

[tool.coverage.run]
source = ["src/maple"]
branch = true
parallel = true
omit = [
    "*/tests/*",
    "*/__init__.py",
]

[tool.coverage.report]
exclude_lines = [
    "pragma: no cover",
    "def __repr__",
    "if self.debug:",
    "if settings.DEBUG",
    "raise AssertionError",
    "raise NotImplementedError",
    "if __name__ == .__main__.:",
    "if TYPE_CHECKING:",
]

# =============================================================================
# MyPy - Static type checking
# =============================================================================

[tool.mypy]
python_version = "3.10"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
disallow_incomplete_defs = true
check_untyped_defs = true
disallow_untyped_decorators = false  # Relaxed for decorators
no_implicit_optional = true
warn_redundant_casts = true
warn_unused_ignores = true
warn_no_return = true
warn_unreachable = true
strict_equality = true
show_error_codes = true

[[tool.mypy.overrides]]
module = [
    "numpy.*",
    "scipy.*",
    "matplotlib.*",
    "pandas.*",
    "networkx.*",
    "sklearn.*",
    "torch.*",
    "pyro.*",
    "seaborn.*",
    "adjustText.*",
    "tqdm.*",
]
ignore_missing_imports = true
